+++ lib/rbs/merge/file_analysis.rb
@@ -145,7 +145,7 @@ module Rbs
       def comment_augmenter(owners: nil, **options)
         comment_tracker.augment(
           owners: owners || comment_augmenter_default_owners,
-          **options
+          **options,
         )
       end